backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
google-generativeai==0.3.2
openpyxl==3.1.5
python-dotenv==1.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.23.0
PyJWT>=2.8.0
